LayoutTests/imported/w3c/IndexedDB-private-browsing/idbobjectstore_add8-expected.txt


PASS IDBObjectStore.add() - object store has autoIncrement:true and the key path is an object attribute
